+++ src/imcntr/__init__.py
@@ -1,6 +1,6 @@
 from .imcntr_utils import Observer
 from .imcntr_connection import SerialCommunication
-from .imcntr_communication import MessageExchange, WaitForMessage, SendMessage
+from .imcntr_communication import MessageExchange, WaitForResponse, GiveOrder
 from .imcntr import Controller, Sample, Shutter, Ready, Connected, Out, In, Clockwise, CounterClockwise, Open, Close, StopMove, StopRotate, Stop
 from importlib.metadata import version