+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 <!--next-version-placeholder-->
 
+## v0.2.0 (15/01/2025)
+
+### Feature
+
+- Add possibility to send instance order or order given as argument when calling instance of `GiveOrder` class.
+
+- Remove `_Advanced_Command` and `_Advanced_Wait` class for better readability.
+
+### Documentation
+
+- Add documentation for new Feature.
+- Remove obsolete documentation.
+- Fixed some typos in docstrings.
+
+
 ## v0.1.0 (19/11/2024)
 
 - First release of `imcntr`!